requirements.txt
pyyaml
requests
dotenv
azure-iot-device
pymongo[srv]
pandas
azure-storage-blob
